requirements.txt
streamlit
pytest
requests
python-dotenv
openai